Begin drags directly on the split screen handle element.
I found that on 12.9" iPads, dragging starting from 99.9% to the right of the window was causing a problem where any open popovers would be dismissed during the transition. I think it was picking up on the tap within the app window, which would dismiss the popover.

With this approach, we actually tap directly on the handle that divides the screen, which is exposed through springboard.